NXT-210 Move logging test into forms dir
