+++ web/src/pages/status.page.tsx
@@ -1,7 +1,5 @@
 import { NextSeo } from "next-seo";
 
-import { logger } from "@/logger";
-
 /**
  * Static status page, used for load balancer (e.g. Varnish) health checks ('probes' in Varnish speak).
  *
@@ -9,10 +7,6 @@ import { logger } from "@/logger";
  * as fast as possible, hence being static.
  */
 export default function StatusPage(): JSX.Element {
-	logger.info("Client side test log (info)");
-	logger.warn("Client side test log (warn)");
-	logger.error("Client side test log (error)");
-
 	return (
 		<>
 			<NextSeo title="Status" noindex />
@@ -20,16 +14,4 @@ export default function StatusPage(): JSX.Element {
 			<p>This is a static status page, used for health checks</p>
 		</>
 	);
-}
-
-export function getServerSideProps() {
-	logger.info("Server side test log (info)");
-	logger.warn("Server side test log (warn)");
-	logger.error("Server side test log (error)");
-
-	return {
-		props: {
-			this: "that",
-		},
-	};
 }
